common/state.c + server/sstate.c + *.h: aux field is a "long" so getaux() should return long

+++ common/state.c
@@ -362,7 +362,7 @@ int state_getflags(st_tree_t *root, const char *var)
 	return sttmp->flags;
 }
 
-int state_getaux(st_tree_t *root, const char *var)
+long state_getaux(st_tree_t *root, const char *var)
 {
 	st_tree_t	*sttmp;
